requirements.txt
anyio==3.6.1
asyncpg==0.26.0
click==8.1.3
fastapi==0.80.0
h11==0.13.0
idna==3.3
loguru==0.6.0
progress==1.6
psycopg2-binary==2.9.3
pydantic==1.9.2
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
